resize and draw commands
Everything seems to be working as there are no errors on resize and
the vertex and color data all seems to be getting to shaders.
Unfortunately there is nothing appearing on the screen but that will all
be sorted out tomorrow I am sure. As I say, all the data looks good.